% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{aktar}
\alias{aktar}
\title{Verileri Kaydetme}
\usage{
aktar(format = NULL, dosyaYolu = NULL)
}
\arguments{
\item{format}{Opsiyonel. Varsayilan formati .csv olarak girilmistir.}

\item{dosyaYolu}{Opsiyonel. Degistirilmezse varsayilan dosya dizinini alir.
Dosya dizinini ogrenmek icin getwd() fonksiyonu kullanilabilir.}
}
\description{
Verilerin .csv, .xlsx veya .txt formatlarinda kaydedilmesini saglayan fonksiyondur.
}
\examples{
\dontrun{aktar(format = ".xlsx", dosyaYolu = "C:/Users/...")}
}
