% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{anahtar}
\alias{anahtar}
\title{API Anahtarını Tanıtma}
\usage{
anahtar(api_anahtar)
}
\arguments{
\item{api_anahtar}{API anahtarı.}
}
\description{
TCMB/EVDS'den alınan API anahtarını tanımlamaya yarayan fonksiyondur.
Anahtar olmadan herhangi bir veri çekme işlemi gerçekleştirilemez.
Yardim için \code{yardim()} fonksiyonu çalıştırılabilir.
}
\examples{
\dontrun{
anahtar(api_anahtar = "api_anahtar")
}
}
