% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_tcmb}
\alias{get_tcmb}
\title{\code{veriler()} için wrapper}
\usage{
get_tcmb(
  series,
  freq = NULL,
  obs = NULL,
  formula = NULL,
  start,
  end = "01-01-3000"
)
}
\arguments{
\item{series}{Seri/serilerin kodu.
Birden fazla kod girmek için \code{c("", "", ...)} şeklinde yazılmalıdır.}

\item{freq}{Opsiyonel.
Detay icin \code{parametreler()} fonksiyonu çalıştırılabilir.
Birden fazla seri kodunda tek bir ortak değer olmalıdır.
Girilmediği zaman sistem serilerin ortak frekansını alacaktır.}

\item{obs}{Opsiyonel.
Detay için \code{parametreler()} fonksiyonu çalıştırılabilir.
Birden fazla seri kodu istenirse gözlem parametreleri her biri için \code{c("", "", ...)} şeklinde tanıtılmalıdır.
Birden fazla seri kodu girilmiş ve tek bir parametre yazılmış ise diğerleri de aynı şekilde gelecektir.
Girilmediği zaman sistem ilgili seri için orijinal gözlemin parametresini uygulayacaktır.}

\item{formula}{Opsiyonel.
Detay için \code{parametreler()} fonksiyonu çalıştırılabilir.
Birden fazla seri kodu istenirse formül parametreleri her biri için \code{c("", "", ...)} seklinde tanıtılmalıdır.
Birden fazla seri kodu girilmiş ve tek bir parametre yazılmış ise diğerleri de ayni sekilde gelecektir.
Girilmediği zaman sistem ilgili seri için düzey parametresini uygulayacaktır.}

\item{start}{\strong{gg-aa-yyyy} formatında girilmelidir.}

\item{end}{Opsiyonel.
\strong{gg-aa-yyyy} formatında girilmelidir.
Girilmezse son tarihi dikkate alacaktır.}
}
\description{
Bir veya daha fazla veriyi indirip analize hazır hale getirir.
\code{veriler()} fonksiyonu için bir wrapper görevi görür.
}
\examples{
\dontrun{get_tcmb(seriler = c(), freq = 5, obs = "avg", formula = 0,
start = "1-1-2009", end = "12-5-2022")}
\dontrun{get_tcmb(seriler = c(), freq = 5, obs = "avg", formula = 0,
start = "1-1-2009", end = "last")}
}
