% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{gorsellestir}
\alias{gorsellestir}
\title{Veri Görselleştirme}
\usage{
gorsellestir(veri, seriKodu, sayisalFrekans)
}
\arguments{
\item{veri}{Veri çerçevesi}

\item{seriKodu}{Görselleştirilmek istenen seri.
NA (kayıp) değerler kaldırılır.}

\item{sayisalFrekans}{Görselleştirilmek istenen serinin sayısal formatta frekansı.
Veri ile ayni frekansta olmalıdır.}
}
\description{
Tek bir seriyi basit bir düzeyde ve interaktif bir sekilde görselleştirmeyi sağlayan fonksiyondur.
Görsel üzerinde yakınlaştırma, kaydetme gibi işlemler yapılabilir.
}
\examples{
\dontrun{gorsellestir(seriKodu = veri$TP_DK_USD_A_YTL, sayisalFrekans = 5)}
}
