% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{kategoriler}
\alias{kategoriler}
\title{Kategorileri Tanıma}
\usage{
kategoriler()
}
\description{
Piyasa Verileri, Kurlar, Faiz İstatistikleri gibi kategorileri içermektedir.
İlgilenilen kategori/kategoriler belirlendikten sonra \code{verigruplari()} fonksiyonunda kullanılacaktır.
}
