% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{parametreler}
\alias{parametreler}
\title{Gözlem, Formül ve Frekans Parametrelerini Tanıma}
\usage{
parametreler()
}
\description{
Veriyi/verileri çekmeden önce filtrelemeyi sağlarlar.
Kullanımı zorunlu değildir.
Belirtilmediği zaman sistem varsayılanı getirecektir.
Bu fonksiyon sadece tanıtım amaçlıdır.
\code{veriler()} fonksiyonunda kullanılacaktır.
Bu fonksiyon üç adet veri çerçevesi üretir: \code{gozlem}, \code{formul}, \code{frekans}.
}
