% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{seriler}
\alias{seriler}
\title{Serileri Tanima}
\usage{
seriler(verigrupKodu)
}
\arguments{
\item{verigrupKodu}{Istenen veri grubunun kodu.}
}
\description{
ABD Dolari (Doviz Alis), ABD Dolari (Doviz Satis) gibi serileri icermektedir.
Ilgilenilen seri/seriler secildikten sonra veriler() fonksiyonunda kullanilacaktir.
}
\examples{
\dontrun{
seriler(verigrupKodu = "bie_dkdovytl")
}
}
