% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{verigruplari}
\alias{verigruplari}
\title{Veri Gruplarını Tanıma}
\usage{
verigruplari(kategoriKodu = NULL, hepsi = FALSE)
}
\arguments{
\item{kategoriKodu}{İstenen kategorinin/kategorilerin kodu. Değer girilecekse hepsi parametresi FALSE bırakılmalıdır. Birden fazla seçilmek istenirse \code{c("", "", ...)} şeklinde girilmelidir.}

\item{hepsi}{Varsayılanı FALSE'tur. Eğer tüm veri grupları istenirse TRUE yapılmalıdır. Bu durumda kategoriKodu parametresine değer girilmemelidir.}
}
\description{
Döviz Kurları, Dış Ticaret gibi veri gruplarını içermektedir.
İlgilenilen veri grubu/grupları seçildikten sonra \code{seriler()} fonksiyonunda kullanılacaktır.
}
\examples{
\dontrun{
verigruplari(kategoriKodu = 2)
verigruplari(kategoriKodu = c(2,14))
verigruplari(hepsi = TRUE)
}
}
