% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{veriler}
\alias{veriler}
\title{Verileri Cekme}
\usage{
veriler(
  seriKodu,
  baslangicTarihi,
  bitisTarihi = "01-01-3000",
  gozlemParametresi = NULL,
  formulParametresi = NULL,
  sayisalFrekans = NULL
)
}
\arguments{
\item{seriKodu}{Veri/verilerin seri kodu.
Birden fazla kod girmek icin c("","",...) seklinde yazilmalidir.}

\item{baslangicTarihi}{gg-aa-yyyy formatinda girilmelidir.}

\item{bitisTarihi}{Opsiyonel.
gg-aa-yyyy formatinda girilmelidir.
Girilmezse son tarihi dikkate alacaktir.}

\item{gozlemParametresi}{Opsiyonel.
Detay icin parametreler() fonksiyonu calistirilabilir.
Birden fazla seri kodu istenirse gozlem parametreleri her biri icin c("","",...) seklinde tanitilmalidir.
Birden fazla seri kodu girilmis ve tek bir parametre yazilmis ise digerleri de ayni sekilde gelecektir.
Girilmedigi zaman sistem ilgili seri icin orijinal gozlemin parametresini uygulayacaktir.}

\item{formulParametresi}{Opsiyonel.
Detay icin parametreler() fonksiyonu calistirilabilir.
Birden fazla seri kodu istenirse formul parametreleri her biri icin c("","",...) seklinde tanitilmalidir.
Birden fazla seri kodu girilmis ve tek bir parametre yazilmis ise digerleri de ayni sekilde gelecektir.
Girilmedigi zaman sistem ilgili seri icin duzey parametresini uygulayacaktir.}

\item{sayisalFrekans}{Opsiyonel.
Detay icin parametreler() fonksiyonu calistirilabilir.
Birden fazla seri kodunda tek bir ortak deger olmalidir.
Girilmedigi zaman sistem serilerin ortak frekansini alacaktir.}
}
\description{
Istenen verinin/verilerin seri kodunun/kodlarinin tanimlanacagi fonksiyondur.
}
\examples{
\dontrun{veriler(seriKodu = "TP.DK.USD.A.YTL", baslangicTarihi = "01-01-2020")}
\dontrun{veriler(seriKodu = "TP.DK.USD.A.YTL", baslangicTarihi = "01-01-2020", bitisTarihi = "01-03-2021", gozlemParametresi = "avg", formulParametresi = 1, sayisalFrekans = 5)}
\dontrun{veriler(seriKodu = c("TP.DK.USD.A.YTL","TP.DK.USD.S.YTL"), baslangicTarihi = "01-01-2020", bitisTarihi = "01-03-2021", gozlemParametresi = "avg", formulParametresi = 1, sayisalFrekans = 5)}
\dontrun{veriler(seriKodu = c("TP.DK.USD.A.YTL","TP.DK.USD.S.YTL"), baslangicTarihi = "01-01-2010", bitisTarihi = "01-12-2020", gozlemParametresi = c("avg","max"), formulParametresi = 1, sayisalFrekans = 8)}
}
