% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{yardim}
\alias{yardim}
\title{API Anahtarının Nasıl Alınacağı Konusunda Yardım}
\usage{
yardim()
}
\description{
Adım adım API anahtarının nasıl alınacağını gösteren fonksiyondur.
Açıklamada bulunan url:
\url{https://evds2.tcmb.gov.tr/index.php?/evds/login}
}
