% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{anahtar}
\alias{anahtar}
\title{API Anahtarini Tanitma}
\usage{
anahtar(api_anahtar)
}
\arguments{
\item{api_anahtar}{API anahtari.}
}
\description{
TCMB/EVDS'den alinan API anahtarini tanimlamaya yarayan fonksiyondur.
Anahtar olmadan herhangi bir veri cekme islemi gerceklestirilemez.
Yardim icin yardim() fonksiyonu calistirilabilir.
}
\examples{
\dontrun{anahtar(api_anahtar = "api_anahtar")}
}
