% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{gorsellestir}
\alias{gorsellestir}
\title{Veri Gorsellestirme}
\usage{
gorsellestir(seriKodu, sayisalFrekans)
}
\arguments{
\item{seriKodu}{Gorsellestirilmek istenen seri.
NA (kayip) degerler kaldirilir.}

\item{sayisalFrekans}{Gorsellestirilmek istenen serinin sayisal formatta frekansi.
Veri ile ayni frekansta olmalidir.}
}
\description{
Tek bir seriyi basit bir duzeyde ve interaktif bir sekilde gorsellestirmeyi saglayan fonksiyondur.
Gorsel uzerinde yakinlastirma, kaydetme gibi islemler yapilabilir.
}
\examples{
\dontrun{gorsellestir(seriKodu = veri$TP_DK_USD_A_YTL, sayisalFrekans = 5)}
}
