% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{kategoriler}
\alias{kategoriler}
\title{Kategorileri Tanima}
\usage{
kategoriler()
}
\description{
Piyasa Verileri, Kurlar, Faiz Istatistikleri gibi kategorileri icermektedir.
Ilgilenilen kategori/kategoriler belirlendikten sonra verigruplari() fonksiyonunda kullanilacaktir.
}
