% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{parametreler}
\alias{parametreler}
\title{Gozlem, Formul ve Frekans Parametrelerini Tanima}
\usage{
parametreler()
}
\description{
Veriyi/verileri cekmeden once filtrelemeyi saglarlar.
Kullanimi zorunlu degildir.
Belirtilmedigi zaman sistem varsayilani getirecektir.
Bu fonksiyon sadece tanitim amaclidir.
veriler() fonksiyonunda kullanilacaktir.
}
