% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{verigruplari}
\alias{verigruplari}
\title{Veri Gruplarini Tanima}
\usage{
verigruplari(kategoriKodu = NULL, hepsi = FALSE)
}
\arguments{
\item{kategoriKodu}{Istenen kategorinin/kategorilerin kodu. Deger girilecekse hepsi parametresi FALSE birakilmalidir. Birden fazla secilmek istenirse c("","",...) seklinde girilmelidir.}

\item{hepsi}{Varsayilani FALSE'tur. Eger tum veri gruplari istenirse TRUE yapilmalidir. Bu durumda kategoriKodu parametresine deger girilmemelidir.}
}
\description{
Doviz Kurlari, Efektif Kurlar gibi veri gruplarini icermektedir.
Ilgilenilen veri grubu/gruplari secildikten sonra seriler() fonksiyonunda kullanilacaktir.
}
\examples{
\dontrun{verigruplari(kategoriKodu = 2)}
\dontrun{verigruplari(kategoriKodu = c(2,14))}
\dontrun{verigruplari(hepsi = TRUE)}
}
