% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{yardim}
\alias{yardim}
\title{API Anahtarinin Nasil Alinacagi Konusunda Yardim}
\usage{
yardim()
}
\description{
Adim adim API anahtarinin nasil alinacagini gosteren fonksiyondur.
Aciklamada bulunan url:
\url{https://evds2.tcmb.gov.tr/index.php?/evds/login}
}
